+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Add `enable_lightcone` option in circuit `expectation` method, where only gates within casual lightcone of local observable is contracted.
 
+- Add `benchmark` function into utils
+
 ### Fixed
 
 - Fixed a vital bug on circuit expectation evaluation, a wrongly transposed operator connection is fixed.